# perf(UpdateGameAchievementsMetrics): reduce query count