`UpdateGameMetrics` tends to be a long-running job that can cause high server load: https://discord.com/channels/476211979464343552/1026595325038833725/1298616276893040722.

Most of the processing time actually takes place in `UpdateGameAchievementsMetrics`, which it calls synchronously:
```php
$game->save();

app()->make(UpdateGameAchievementsMetrics::class)
    ->execute($game);
```

The main bottleneck in `UpdateGameAchievementsMetrics` is `$unlocksCount` and `$unlocksHardcoreCount` are both fetched in a loop for every achievement in the set.

This PR optimizes the action by lifting this up into a single query.

**How to test this PR:**
In _app/Platform/Commands/UpdateGameMetrics.php_, put this at the top of `handle()`:
```php
$startTime = microtime(true);
```
And put this at the bottom of `handle()`:
```php
$endTime = microtime(true);
$executionTime = $endTime - $startTime;
$this->line("Command executed in " . round($executionTime, 2) . " seconds.");
```
Then run:
```bash
sail artisan ra:platform:game:update-metrics 228
```

**UpdateGameMetrics Execution Time Before**
1.65s
1.61s
1.62s
1.66s
1.74s
1.62s
Average: 1.65s

**UpdateGameMetrics Execution Time After**
1.12s
1.16s
1.21s
1.25s
1.14s
1.18s
Average: 1.18s (~28% improvement)